backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg>=0.30.0
sqlalchemy[asyncio]==2.0.36
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
beautifulsoup4==4.12.3
ruff>=0.6.0
